--- a/src/Modules/UngradedModules/FreeModuleHom.jl
+++ b/src/Modules/UngradedModules/FreeModuleHom.jl
@@ -559,9 +559,11 @@ Homogeneous module homomorphism)
 ```
 """
 function image(h::FreeModuleHom)
-  si = filter(x->!iszero(x), images_of_generators(h))
+  si = filter(!iszero, copy(images_of_generators(h)))
   s = sub(codomain(h), si, :module)
-  return s, hom(s, codomain(h), si, check=false)
+  phi = hom(s, codomain(h), si, check=false)
+  phi.generators_map_to_generators = true
+  return s, phi
 end