… src/Modules/UngradedModules/SubModuleOfFreeModule.jl
… src/Modules/UngradedModules/SubQuoHom.jl
… src/Modules/UngradedModules/SubquoModuleElem.jl
… src/Modules/UngradedModules/Tensor.jl