@jankoboehm : I wanted to look a bit into the morphisms of modules for the reasons we talked about today. But before I could actually start doing that, I needed to split that gigantic file into more digestable parts. For me personally this was so that syntax highlighting would not crash all the time. But also in general, I think, it makes sense to split this into thematically ordered smaller chunks. 

Do you agree?