@HechtiDerLachs 2023-11-23
I tried to improve the code a bit, especially regarding the speed of evaluation of mappings. So far running the same test-suite for modules gives 

before:
```
 45.207242 seconds (433.82 M allocations: 20.793 GiB, 12.69% gc time, 17.17% compilation time)
 45.368074 seconds (433.65 M allocations: 20.792 GiB, 13.04% gc time, 17.48% compilation time)
```
after
```
 37.961641 seconds (265.70 M allocations: 12.739 GiB, 10.21% gc time, 21.95% compilation time)
 38.276709 seconds (266.08 M allocations: 12.747 GiB, 7.61% gc time, 22.06% compilation time)
 37.249246 seconds (265.81 M allocations: 12.744 GiB, 7.62% gc time, 21.19% compilation time)
```
So especially in terms of memory consumption, I think this is useful and should be merged one way or the other.