@thofma 2023-11-23
Why not separate the code reorganization from the improvements PR wise? The spitting should definitely be happening.