@HechtiDerLachs 2023-11-24
I ran some more tests locally. Compared to current master I now get the following timings:

before:
```
46.521581 seconds (434.29 M allocations: 20.809 GiB, 12.73% gc time, 16.09% compilation time)
 47.341869 seconds (433.96 M allocations: 20.798 GiB, 9.42% gc time, 16.86% compilation time)
```
after:
```
 39.683286 seconds (273.27 M allocations: 12.962 GiB, 10.36% gc time, 21.05% compilation time)
 38.410556 seconds (273.16 M allocations: 12.958 GiB, 7.44% gc time, 21.16% compilation time)
 39.105477 seconds (273.33 M allocations: 12.962 GiB, 7.44% gc time, 20.39% compilation time)
```
The difference to the above runs is due to some tests being disabled (on both sides) before. These runs here really take the whole test suite for the modules.