@HechtiDerLachs 2023-11-24
@jankoboehm : I double checked again. There is no internal constructor for neither `FreeModuleHom` nor `SubQuoHom` which does not take the images of the generators as input. So computing and caching them only on request has never been an option.