@HechtiDerLachs 2023-11-24
Timings now go down to
```
 38.644607 seconds (268.27 M allocations: 12.779 GiB, 7.30% gc time, 20.29% compilation time)
 38.324128 seconds (268.55 M allocations: 12.780 GiB, 7.56% gc time, 21.69% compilation time)
```
No major improvement anymore. But still something. I think I can stop here. 

Could maybe someone confirm the changes with the hashing?